Derive coproduct Picklers with macros
