+++ core/src/main/scala-3/sttp/tapir/macros/ValidatorMacros.scala
@@ -26,7 +26,7 @@ private[tapir] object ValidatorMacros {
       report.errorAndAbort("Can only enumerate values of a sealed trait, class or enum.")
     }
 
-    val instances = enumerationTypeChildren[T](failOnError = true).distinct
+    val instances = enumerationTypeChildren[T](failOnError = true).flatMap(_.toList).distinct
       .sortBy(_.name)
       .map(x =>
         tpe.memberType(x).asType match {
