+++ doc/endpoint/pickler.md
@@ -87,9 +87,10 @@ import sttp.tapir.generic.Configuration
 given customConfiguration: Configuration = Configuration.default.withSnakeCaseMemberNames
 ```
 
-## Enums / sealed traits / coproducts
+## Sealed traits / coproducts
 
-Pickler derivation for coproduct types (enums / sealed hierarchies) works automatically, by adding an `$type` discriminator field with the full class name. This is the default behavior of uPickle, but it can be overridden either by changing the discriminator field name, or by using custom logic to get field value from base trait.
+Pickler derivation for coproduct types (enums with parameters / sealed hierarchies) works automatically, by adding a `$type` discriminator field with the full class name. This is the default behavior of uPickle, but it can be overridden either by changing the discriminator field name, or by using custom logic to get field value from base trait.
+Selaed hierarchies with all cases being objects are treated differently, considered as [enumerations](#enumerations).
 
 A discriminator field can be specified for coproducts by providing it in the configuration; this will be only used during automatic and semi-automatic derivation:
 
@@ -136,13 +137,22 @@ Schemas generated by picklers can be customized using annotations, just like wit
 
 ## Enumerations
 
-Scala 3 `enums`, where all cases are parameterless, are treated as an enumeration (not as a coproduct / sealed hierarchy). They are also automatically handled by `Pickler.derived[T]`: enum values are encoded as simple strings representing the type name. For example:
+Tapir schemas and JSON codecs treats following cases as "enumerations":
+1. Scala 3 `enums`, where all cases are parameterless
+2. Sealed hierarchies (coproducts), where all cases are case objects
+
+Such types are handled by `Pickler.derived[T]`: possible values are encoded as simple strings representing the case objects. For example:
 
 ```scala 
 import sttp.tapir.json.pickler.*
 
 enum ColorEnum:
   case Green, Pink
+
+// or:
+// sealed trait ColorEnum
+//   case object Green extends ColorEnum
+//   case object Pink extends ColorEnum
 
 case class ColorResponse(color: ColorEnum, description: String)
 
@@ -157,7 +167,27 @@ pResponse.toCodec.encode(
 pResponse.schema
 ```
 
-If you need to customize enum value encoding, use `Pickler.derivedEnumeration[T]`:
+If sealed hierarchy or enum contain case classes with parameters, they are no longer an "enumeration", and will be treated as standard sealed hierarchies (coproducts):
+
+```scala
+import sttp.tapir.json.pickler.*
+
+sealed trait ColorEnum
+case object Green extends ColorEnum
+case class Pink(intensity: Int) extends ColorEnum
+
+case class ColorResponse(color1: ColorEnum, color2: ColorEnum)
+
+given Pickler[ColorEnum] = Pickler.derived
+val pResponse = Pickler.derived[ColorResponse]
+
+// {"color1":{"$type":"my.app.Pink","intensity":85},"color2":{"$type":"my.app.Green"}}
+pResponse.toCodec.encode(
+  ColorResponse(Pink(85), Green)
+)
+```
+
+If you need to customize enumeration value encoding, use `Pickler.derivedEnumeration[T]`:
 
 ```scala 
 import sttp.tapir.json.pickler.*
@@ -191,4 +221,6 @@ you can proceed with `Pickler.derived[T]`.
 
 * Tapir pickler serialises None values as `null`, instead of wrapping the value in an array
 * Value classes (case classes extending AnyVal) will be serialised as simple values
+
+
