+++ json/pickler/src/main/scala/sttp/tapir/json/pickler/SubtypeDiscriminator.scala
@@ -18,10 +18,4 @@ private[pickler] trait CustomSubtypeDiscriminator[T] extends SubtypeDiscriminato
   // to integrate with uPickle where at some point all we have is Any
   def writeUnsafe(t: Any): String = asString(extractor(t.asInstanceOf[T]))
 
-/** Describes non-standard encoding/decoding and validation for enums. Allows specifying an encoder function which transforms an enum value
-  * to String for serialization (for example by referring a field in enum's base trait, or calling .ordinal.toString for numbers).
-  */
-private[pickler] case class EnumValueDiscriminator[T](encode: T => String, validator: Validator.Enumeration[T])
-    extends SubtypeDiscriminator[T]
-
 private[pickler] case class DefaultSubtypeDiscriminator[T]() extends SubtypeDiscriminator[T]
